+++ server/src/http_objects.rs
@@ -208,8 +208,8 @@ impl From<data_model::Node> for Node {
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct RuntimeInformation {
-    pub major_version: u32,
-    pub minor_version: u32,
+    pub major_version: u8,
+    pub minor_version: u8,
 }
 
 impl From<RuntimeInformation> for data_model::RuntimeInformation {
